[WL Token] TAMUcoin - Tradable Access Monetary Units
Project Name
Tradable Access Monetary Units

Description of the project
Tradable Access Monetary Units seek to enable cross border exchanges of intrinsic value by allowing holders to offer information, service or goods in exchange for TAMUcoins. The concept allows people to offer their expertise to a larger group that may benefit from unique knowledge or skills without currency exchange and security issues. Holders are also entitled to join chat communities designed to spread the DeFi message and help holders grow in their understanding of decentralized finance and the blockchain while discovering new projects reviewed by other holders and discussed freely.

Website or public page of the project
tamucoin.io

What is the name of the token you are adding?
TAMU is the symbol, the coins are "Tradable Access Monetary Units"

Notes or questions to developers [optional]
We were instructed to add our coin to the whitelist by Makoto on Telegram via a private conversation where he asked me to share our vision and website with him directly. The icon file is attached, but please inform us if the format is incorrect or if additional steps are required.