Added UMI token to whitelist.json
Project Name
Umi Digital

Description of the project
Umi is a decentralised art project that combines DeFi farming with NFTs in a modern digital art gallery experience.
Users can farm, buy, sell and mint their own NFTs. 
Umi token is a utility token for Umi.Digital. It is used to pay for services or the purchase of digital content in the Umi Digital ecosystem. Equally as important, it can be staked to farm rare NFTs and earn staking rewards.

Website or public page of the project
Homepage: https://umi.digital/
Token Page: https://umi.digital/umi-token/
Contact Email: contact@umi.digital

What is the name of the token you are adding?
UmiToken